+++ lib/codelog/command/step/changelog.rb
@@ -31,11 +31,8 @@ module Codelog
 
         def create_file_from(changes)
           File.open('CHANGELOG.md', 'w+') do |f|
-            f.puts %(# Changelog
-  All notable changes to this project will be documented in this file.
-  The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
-  and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
-            )
+            f.puts "# Changelog"
+            f.puts Codelog::Config.header
             f.puts(changes)
           end
         end