+++ CHANGELOG.md
@@ -1,8 +1,30 @@
 # Changelog
-  All notable changes to this project will be documented in this file.
-  The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
-  and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
-            
+All notable changes to this project will be documented in this file.
+ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
+ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+## [v0.4]
+### Added
+- New features implemented
+
+### Changed
+- Changes to existing feature
+
+### Deprecated
+- Features that are soon to be removed
+
+### Removed
+- Features that were removed
+
+### Fixed
+- Changes to broken code
+
+### Security
+- Changes that fix vulnerabilities
+
+### Deploy notes
+- Changes that should impact the deploy process and what should be made before it
+
+---
 ## 0.3.0 - 2018-01-18
 ### Added
 - Datestamps along with version tag (synthax codelog release [x.y.z] <DATE>)